refact(goodbits): modularise functions for better testability
